+++ packages/expo-camera/CHANGELOG.md
@@ -6,11 +6,15 @@
 
 ### 🎉 New features
 
-- Add `autoFocus` prop to allow setting the device focus mode. ([#28650](https://github.com/expo/expo/pull/28650) by [@alanjhughes](https://github.com/alanjhughes))
-
 ### 🐛 Bug fixes
 
 ### 💡 Others
+
+## 15.0.4 — 2024-05-07
+
+### 🎉 New features
+
+- Add `autoFocus` prop to allow setting the device focus mode. ([#28650](https://github.com/expo/expo/pull/28650) by [@alanjhughes](https://github.com/alanjhughes))
 
 ## 15.0.3 — 2024-04-29